+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Not released yet
 ## Popup
 - Added a delete option to the blacklisted sites list within settings for every blacklisted site
+- Added the number of blacklisted sites to the button that feeds the list of blacklisted sites
 
 ---